+++ contracts/interfaces/IChai.sol
@@ -6,7 +6,7 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity >=0.5.0 <0.6.0;
 
-import '@openzeppelin-2.5.0/token/ERC20/IERC20.sol';
+import "@openzeppelin-2.5.0/token/ERC20/IERC20.sol";
 
 interface IPot {
   function dsr() external view returns (uint256);
@@ -17,11 +17,15 @@ interface IPot {
 }
 
 contract IChai is IERC20 {
-  function move(address src, address dst, uint256 wad) external returns (bool);
+  function move(
+    address src,
+    address dst,
+    uint256 wad
+  ) external returns (bool);
 
   function join(address dst, uint256 wad) external;
 
   function draw(address src, uint256 wad) external;
 
-  function exit(address src, uint wad) external;
+  function exit(address src, uint256 wad) external;
 }
