+++ contracts/interfaces/IPancakeFactory.sol
@@ -7,7 +7,7 @@
 pragma solidity >=0.5.0 <0.6.0;
 
 interface IPancakeFactory {
-  event PairCreated(address indexed token0, address indexed token1, address pair, uint);
+  event PairCreated(address indexed token0, address indexed token1, address pair, uint256);
 
   function feeTo() external view returns (address);
 
@@ -15,9 +15,9 @@ interface IPancakeFactory {
 
   function getPair(address tokenA, address tokenB) external view returns (address pair);
 
-  function allPairs(uint) external view returns (address pair);
+  function allPairs(uint256) external view returns (address pair);
 
-  function allPairsLength() external view returns (uint);
+  function allPairsLength() external view returns (uint256);
 
   function createPair(address tokenA, address tokenB) external returns (address pair);
