+++ contracts/stakingv2/delegation/VoteDelegatorState.sol
@@ -6,10 +6,10 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.0;
 
-import '@openzeppelin-4.8.0/token/ERC20/utils/SafeERC20.sol';
-import './VoteDelegationUpgradeable.sol';
-import '../../mixins/EnumerableBytes32Set.sol';
-import '../../../interfaces/IStakingV2.sol';
+import "@openzeppelin-4.8.0/token/ERC20/utils/SafeERC20.sol";
+import "./VoteDelegationUpgradeable.sol";
+import "../../mixins/EnumerableBytes32Set.sol";
+import "../../../interfaces/IStakingV2.sol";
 
 contract VoteDelegatorState is VoteDelegationUpgradeable {
   // A record of each accounts delegate
@@ -29,13 +29,13 @@ contract VoteDelegatorState is VoteDelegationUpgradeable {
   mapping(address => uint32) public numCheckpoints;
 
   /// @notice The EIP-712 typehash for the contract's domain
-  bytes32 public constant DOMAIN_TYPEHASH = keccak256('EIP712Domain(string name,uint256 chainId,address verifyingContract)');
+  bytes32 public constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
 
   /// @notice The EIP-712 typehash for the delegation struct used by the contract
-  bytes32 public constant DELEGATION_TYPEHASH = keccak256('Delegation(address delegatee,uint256 nonce,uint256 expiry)');
+  bytes32 public constant DELEGATION_TYPEHASH = keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
 
   /// @notice A record of states for signing / validating signatures
-  mapping(address => uint) public nonces;
+  mapping(address => uint256) public nonces;
 
   mapping(address => uint256) public totalDelegators;
